Fix Module Option Require

+++ test/AddModuleTest.php
@@ -4,7 +4,7 @@ use Tests\TestCase;
 
 require_once("TestUtils.php");
 
-class CreateModuleTest extends TestCase
+class AddModuleTest extends TestCase
 {
 
     private $workdir = __DIR__ . "/files";
@@ -14,7 +14,8 @@ class CreateModuleTest extends TestCase
      */
     public function tearDown() : void {
         $utils = new TestUtils();
-        $utils->deleteDir($this->workdir . "/modules");
+        if(file_exists($this->workdir . "/modules"))
+            $utils->deleteDir($this->workdir . "/modules");
     }
 
     /**
@@ -22,7 +23,7 @@ class CreateModuleTest extends TestCase
      */
     public function it_can_create_a_module()
     {
-        $this->artisan("module:make", [
+        $this->artisan("module:add", [
             "name" => "Test"
         ])->expectsOutput("Module Test Created");
 
@@ -34,15 +35,15 @@ class CreateModuleTest extends TestCase
      */
     public function it_create_the_provider_for_module()
     {
-        $this->artisan("module:make", [
+        $this->artisan("module:add", [
             "name" => "Test"
         ])
         ->expectsOutput("'providers' => [
             ...
-            Test\\Providers\\TestProvider::class
+            Test\\Provider\\TestProvider::class
         ];");
 
-        $this->assertFileExists($this->workdir . "/modules/Test/Providers/TestProvider.php");
+        $this->assertFileExists($this->workdir . "/modules/Test/Provider/TestProvider.php");
     }
 
     /**
@@ -50,7 +51,7 @@ class CreateModuleTest extends TestCase
      */
     public function it_can_register_the_new_module_in_composer_json()
     {
-        $this->artisan("module:make", [
+        $this->artisan("module:add", [
             "name" => "Test"
         ]);
 
@@ -62,13 +63,13 @@ class CreateModuleTest extends TestCase
      */
     public function it_can_create_a_module_with_sub_folders()
     {
-        $this->artisan("module:make", [
+        $this->artisan("module:add", [
             "name" => "User/Test"
         ]);
 
         $this->assertStringContainsString("User\\\\Test\\\\", file_get_contents($this->workdir."/composer.json"));
-        $this->assertFileExists($this->workdir . "/modules/User/Test/Providers/TestProvider.php");
-        $this->assertStringContainsString("namespace User\Test\Providers", file_get_contents($this->workdir . "/modules/User/Test/Providers/TestProvider.php"));
+        $this->assertFileExists($this->workdir . "/modules/User/Test/Provider/TestProvider.php");
+        $this->assertStringContainsString("namespace User\Test\Provider", file_get_contents($this->workdir . "/modules/User/Test/Provider/TestProvider.php"));
     }
 
     /**
@@ -76,13 +77,13 @@ class CreateModuleTest extends TestCase
      */
     public function it_can_create_a_module_with_many_sub_folders()
     {
-        $this->artisan("module:make", [
+        $this->artisan("module:add", [
             "name" => "Group/User/Test"
         ]);
 
         $this->assertStringContainsString("Group\\\\User\\\\Test\\\\", file_get_contents($this->workdir."/composer.json"));
-        $this->assertFileExists($this->workdir . "/modules/Group/User/Test/Providers/TestProvider.php");
-        $this->assertStringContainsString("namespace Group\User\Test\Providers", file_get_contents($this->workdir . "/modules/Group/User/Test/Providers/TestProvider.php"));
+        $this->assertFileExists($this->workdir . "/modules/Group/User/Test/Provider/TestProvider.php");
+        $this->assertStringContainsString("namespace Group\User\Test\Provider", file_get_contents($this->workdir . "/modules/Group/User/Test/Provider/TestProvider.php"));
     }
 
 }
