+++ test/MakeModelForModuleTest.php
@@ -4,9 +4,10 @@ namespace ModuleGenerator\Test;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
+use ModuleGenerator\Commands\Constants;
 use Tests\TestCase;
 
-class ModuleModelTest extends TestCase
+class MakeModelForModuleTest extends TestCase
 {
     private $workdir = __DIR__ . "/files";
 
@@ -15,7 +16,8 @@ class ModuleModelTest extends TestCase
      */
     public function tearDown() : void {
         $utils = new TestUtils();
-        $utils->deleteDir($this->workdir . "/modules");
+        if(file_exists($this->workdir . "/modules"))
+            $utils->deleteDir($this->workdir . "/modules");
     }
 
     /**
@@ -46,5 +48,20 @@ class ModuleModelTest extends TestCase
         ])->expectsOutput("TestModel model created");
 
         $this->assertFileExists($this->workdir . "/modules/Test/Auth/Model/TestModel.php");
+    }
+
+    /**
+     * A basic feature test example.
+     * @test
+     * @return void
+     *
+     */
+    public function module_option_is_required()
+    {
+        $this->artisan("module:model", [
+            "name" => "TestModel",
+        ])->expectsOutput(Constants::MODULE_OPTION_REQUIRED);
+
+        $this->assertFileDoesNotExist($this->workdir . "/modules/Test/Auth/Model/TestModel.php");
     }
 }
