+++ test/MakeRoutesForModuleTest.php
@@ -1,9 +1,11 @@
 <?php
 
 namespace ModuleGenerator\Test;
+use ModuleGenerator\Commands\Constants;
+use ModuleGenerator\Commands\MakeRoutesForModule;
 use Tests\TestCase;
 
-class RoutesCommandTest extends TestCase
+class MakeRoutesForModuleTest extends TestCase
 {
 
     private $workdir = __DIR__ . "/files";
@@ -13,7 +15,8 @@ class RoutesCommandTest extends TestCase
      */
     public function tearDown() : void {
         $utils = new TestUtils();
-        $utils->deleteDir($this->workdir . "/modules");
+        if(file_exists($this->workdir . "/modules"))
+            $utils->deleteDir($this->workdir . "/modules");
     }
 
     /**
@@ -30,5 +33,20 @@ class RoutesCommandTest extends TestCase
           ->expectsOutput("Add routes in boot module provider method, e.g.: ");
 
           $this->assertFileExists($this->workdir . "/modules/Test/routes.php");
+    }
+
+    /**
+     * A basic feature test example.
+     * @test
+     * @return void
+     *
+     */
+    public function module_option_is_required()
+    {
+        $response = $this->artisan("module:routes",[
+            "name" => "routes",
+        ])->expectsOutput(Constants::MODULE_OPTION_REQUIRED);
+
+        $this->assertFileDoesNotExist($this->workdir . "/modules/routes.php");
     }
 }
